:hammer: Pass data retrieved from DB to ejs
