+++ src/views/index.ejs
@@ -3,12 +3,102 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title><%= text %></title>
+    <title>ProtectMe</title>
+    <link
+      href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.16/dist/tailwind.min.css"
+      rel="stylesheet"
+    />
   </head>
+  <body class="bg-gray-100">
+    <header class="bg-white shadow-md">
+      <div class="container mx-auto px-4 py-4">
+        <div class="flex justify-between items-center">
+          <div class="text-blue-600 font-bold text-xl">
+            <a href="/">ProtectMe</a>
+          </div>
+          <div class="flex space-x-4 items-center">
+            <button class="bg-green-600 text-white rounded-md px-4 py-2">
+              Sign up
+            </button>
+            <button class="bg-blue-600 text-white rounded-md px-4 py-2">
+              Login
+            </button>
+          </div>
+        </div>
+      </div>
+    </header>
 
-  <body>
-    <h1><%= text %></h1>
+    <main class="container mx-auto px-4 py-16">
+      <h1 class="text-4xl font-bold mb-6">Welcome to our Blog</h1>
+      <div class="mb-6">
+        <form>
+          <div class="relative">
+            <input
+              type="text"
+              class="border-2 border-gray-300 rounded-md w-full pl-10 p-2"
+              placeholder="Search posts..."
+            />
+            <svg
+              class="absolute top-1/2 left-3 transform -translate-y-1/2 h-4 w-4 text-gray-400"
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 24 24"
+              fill="none"
+              stroke="currentColor"
+              stroke-width="2"
+              stroke-linecap="round"
+              stroke-linejoin="round"
+            >
+              <circle cx="11" cy="11" r="8"></circle>
+              <line x1="21" y1="21" x2="16.65" y2="16.65"></line>
+            </svg>
+          </div>
+        </form>
+      </div>
+      <% blogs.forEach(function(blog) { %>
+      <div class="divide-y divide-gray-200 mb-3">
+        <a href="/" class="block">
+          <div
+            class="bg-white shadow-md rounded-md p-4 hover:cursor-pointer transition"
+          >
+            <div class="flex items-start">
+              <img
+                src="<%= blog.user.img %>"
+                alt="User Image"
+                class="rounded-full w-10 h-10 mr-4"
+              />
+              <div>
+                <div class="text-gray-700 text-sm">
+                  <span>@<%= blog.user.name %></span>
+                </div>
+                <div class="text-gray-700 text-sm mb-2">
+                  <span><%= blog.createdAt.toISOString().slice(0, 10) %></span>
+                </div>
+                <h2 class="text-xl font-semibold mb-2"><%= blog.title %></h2>
+                <div class="flex items-center">
+                  <div class="flex items-center -ml-1">
+                    <svg
+                      class="h-5 w-5 text-gray-400"
+                      xmlns="http://www.w3.org/2000/svg"
+                      viewBox="0 0 24 24"
+                      fill="none"
+                      stroke="currentColor"
+                      stroke-width="1"
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                    >
+                      <path
+                        d="M12 21C12 21 3 14.9091 3 8.63636C3 5.63636 5.2381 3 8.09524 3C10.3808 3 11.9996 4.61884 12 6.90909C12.0004 4.61884 13.6192 3 15.9048 3C18.7619 3 21 5.63636 21 8.63636C21 14.9091 12 21 12 21Z"
+                      />
+                    </svg>
+                    <span class="ml-1 text-black-400">123</span>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </a>
+      </div>
+      <% }) %>
+    </main>
   </body>
 </html>
-
-:hammer: Print "Hello world!" from ejs
